requirements.txt
# KONACH - Runtime Dependencies

PyQt5>=5.15.9
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0
pymysql>=1.1.0
jinja2>=3.1.0
weasyprint>=61.0
cryptography>=42.0.0
